Feature of changing grid size using slider & bug fixes
Wrap the loops generating the grid, inside a function gridGenerator.
gridGenerator takes value of number of pixels from slider. It is called
when either the page loads or when there is change in slider value

Whenever gridgenerator is called, the entire previous grid is removed
using innerHTML function and the value of penState variable and
clickState is reset and so that any previous pen settings do not
interfere with the newer grid.

Listen mouse up and mouse down event over entire document. Listening
these events just on the grid was causing unwanted results whenever
cursor was moved out and left mouse button was released.